p38  增加了回复/评论的功能

+++ src/main/java/com/thymeleaf01/controller/PublishController.java
@@ -1,8 +1,6 @@
 package com.thymeleaf01.controller;
 
 import com.thymeleaf01.dto.QuestionDTO;
-import com.thymeleaf01.mapper.QuestionMapper;
-import com.thymeleaf01.mapper.UserMapper;
 import com.thymeleaf01.model.Question;
 import com.thymeleaf01.model.User;
 import com.thymeleaf01.service.QuestionService;
@@ -21,7 +19,7 @@ public class PublishController {
     @Resource
     private QuestionService questionService;
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -40,7 +38,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description",required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             HttpServletRequest request,
             Model model){
         model.addAttribute("title",title);
