+++ src/main/java/com/thymeleaf01/service/QuestionService.java
@@ -53,7 +53,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public PaginationDTO list(Integer userId, Integer page, Integer size) {
+    public PaginationDTO list(Long userId, Integer page, Integer size) {
         PaginationDTO paginationDTO = new PaginationDTO();
         Integer totalPage;
         Integer totalCount = questionMapper.countByUserId(userId);
@@ -85,7 +85,7 @@ public class QuestionService {
         return paginationDTO;
     }
 
-    public QuestionDTO getById(Integer id) {
+    public QuestionDTO getById(Long id) {
         QuestionDTO question = questionMapper.getById(id);
         if(question ==null){
             throw new CustomizeException(CustomizeErrorCode.QUESTION_NOT_FOUND);
@@ -98,8 +98,12 @@ public class QuestionService {
     }
     public void createOrUpdate(Question question) {
         if (question.getId() == null){
+            //创建问题
             question.setGmtCreat(System.currentTimeMillis());
             question.setGmtModified(question.getGmtCreat());
+//            question.setViewCount(0);
+//            question.setLikeCount(0);
+//            question.setCommentCount(0);
             int create = questionMapper.create(question);
             System.out.println("新增问题完成，返回create= "+ create);
         }else {
@@ -112,7 +116,7 @@ public class QuestionService {
         }
     }
 
-    public void incView(Integer id) {
+    public void incView(Long id) {
         questionMapper.incView(id);
     }
 }
